@@ -61,26 +64,28 @@
     "@types/fs-extra": "^9.0.1",
     "@types/hapi__joi": "^17.1.4",
     "@types/inquirer": "^7.3.0",
-    "@types/joi": "^14.3.4",
     "@types/js-yaml": "^3.12.5",
     "@types/json5": "0.0.30",
     "@types/luxon": "^1.24.1",
-    "@types/mocha": "^8.0.2",
-    "@types/node": "^14.0.27",
+    "@types/mocha": "^8.0.3",
+    "@types/node": "^14.6.0",
     "@types/nodemon": "^1.19.0",
     "@types/recursive-readdir": "^2.2.0",
     "@types/update-notifier": "^4.1.1",
     "@types/uuid": "^8.3.0",
-    "@types/yeoman-environment": "^2.10.0",
-    "@types/yeoman-generator": "^4.11.1",
+    "@types/yeoman-environment": "^2.10.1",
+    "@types/yeoman-generator": "^4.11.2",
     "@types/yeoman-test": "^2.0.4",
+    "@typescript-eslint/eslint-plugin": "^3.10.1",
+    "@typescript-eslint/parser": "^3.10.1",
     "body-parser": "^1.19.0",
     "chai-as-promised": "^7.1.1",
     "chai-fs": "^2.0.0",
-    "eslint": "^7.6.0",
+    "eslint": "^7.7.0",
     "eslint-config-oclif": "^3.1.0",
     "eslint-config-oclif-typescript": "^0.2.0",
     "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-chai-friendly": "^0.6.0",
     "globby": "^11.0.1",
     "nock": "^13.0.4",
     "npm-check": "^5.9.0",
@@ -88,8 +93,9 @@
     "shx": "^0.3.2",
     "sinon": "^9.0.3",
     "source-map-support": "^0.5.17",
-    "ts-node": "^8.10.2",
-    "typescript": "^3.9.7",
+    "tar": "^6.0.5",
+    "ts-node": "^9.0.0",
+    "typescript": "^4.0.2",
     "yeoman-assert": "^3.1.1",
     "yeoman-test": "^3.0.0"
   },
