@@ -3,9 +3,9 @@ import { flags } from "@oclif/command";
 import { checkAppLoginStatus } from "../core/utils/users";
 
 export default class AppsIndex extends BaseCommand {
-  static description = "list your apps";
+  public static description = "list your apps";
 
-  static flags = {
+  public static flags = {
     help: flags.help({
       char: "h",
       description: "show help for the apps commands",
@@ -13,19 +13,21 @@ export default class AppsIndex extends BaseCommand {
   };
 
   // hide the command from help
-  static hidden = true;
+  public static hidden = true;
 
-  async run() {
+  async run(): Promise<void> {
     // When the -h flag is present the following line haults execution
     this.parse(AppsIndex);
 
     await checkAppLoginStatus(this);
 
     try {
-      const apps = this.appsClient!.apps.getAll();
-      (await (await apps).items).forEach((app) => {
-        this.log(app.name);
-      });
+      if (this.appsClient) {
+        const apps = this.appsClient.apps.getAll();
+        (await apps).items.forEach((app) => {
+          this.log(app.name);
+        });
+      }
     } catch (error) {
       this.error(error);
     }
