@@ -1,8 +1,9 @@
 import BaseCommand from "../base-command";
 import Test from "./test";
-import publishApp from "../core/publish-app";
+import publishApp, { isAppFailedToDeployError } from "../core/publish-app";
 import { flags } from "@oclif/command";
 import { checkAppLoginStatus } from "../core/utils/users";
+import { packageApp, isAppFailedToPackageError } from '../core/package-app';
 
 export default class Publish extends BaseCommand {
   static description = "publish your app";
@@ -16,9 +17,9 @@ export default class Publish extends BaseCommand {
       char: "h",
       description: "show help for the publish command",
     }),
-    watch: flags.boolean({
-      char: "w",
-      description: "check the status of the deployment until complete",
+    "no-watch": flags.boolean({
+      char: "n",
+      description: "does not track the status of the deployment",
     }),
     "skip-tests": flags.boolean({
       char: "s",
@@ -27,7 +28,7 @@ export default class Publish extends BaseCommand {
     }),
   };
 
-  async run() {
+  async run(): Promise<void> {
     // When the -h flag is present the following line haults execution
     const { flags } = this.parse(Publish);
 
@@ -36,23 +37,27 @@ export default class Publish extends BaseCommand {
     if (!flags["skip-tests"]) await Test.run(["-f"]);
 
     try {
-      const pathToApp = process.cwd();
-      await publishApp(pathToApp, this.appsClient!, {
-        watch: flags.watch,
-      });
-    } catch (error) {
-      switch (error.code) {
-        case "APP_FAILED_TO_PACKAGE":
-          return this.error(error.message, {
-            exit: 1,
-          });
-        case "APP_FAILED_TO_DEPLOY":
-          return this.error(error.message, {
-            exit: 1,
-          });
-        default:
-          throw error;
+      const tarballName = await packageApp();
+      if(this.appsClient) {
+        await publishApp(tarballName, this.appsClient, {
+          noWatch: flags["no-watch"],
+        });
       }
+    } catch (error) {
+
+      if (isAppFailedToPackageError(error)) {
+        return this.error(error.message, {
+          exit: 1,
+        });
+      }
+
+      if (isAppFailedToDeployError(error)) {
+        return this.error(error.message, {
+          exit: 1,
+        });
+      }
+
+      throw error;
     }
   }
 }
