@@ -6,6 +6,7 @@ import {
   NetworkErrorCollection,
   PaginatedItems,
 } from "../../types";
+import { AxiosError } from 'axios';
 
 export default class Deployments {
   private client: ShipengineAPIClient;
@@ -37,12 +38,15 @@ export default class Deployments {
           "content-type": `multipart/form-data; boundary=${form.getBoundary()}`,
         },
         isFileUpload: true,
-      });
+      }) as Deployment;
 
       return Promise.resolve(response);
     } catch (error) {
-      return Promise.reject(error.response.data as NetworkErrorCollection);
-    }
+      const err = error as AxiosError;
+      if (err.response) {
+        return Promise.reject(err.response.data as NetworkErrorCollection);
+      }
+      return Promise.reject(err.message);    }
   }
 
   /**
@@ -54,12 +58,15 @@ export default class Deployments {
       const response = await this.client.call({
         endpoint: `apps/${appId}/deploys`,
         method: "GET",
-      });
+      }) as PaginatedItems<Deployment>;
 
       return Promise.resolve(response);
     } catch (error) {
-      return Promise.reject(error.response.data as NetworkErrorCollection);
-    }
+      const err = error as AxiosError;
+      if (err.response) {
+        return Promise.reject(err.response.data as NetworkErrorCollection);
+      }
+      return Promise.reject(err.message);    }
   }
 
   /**
@@ -77,12 +84,15 @@ export default class Deployments {
       const response = await this.client.call({
         endpoint: `apps/${appId}/deploys/${deployId}`,
         method: "GET",
-      });
+      }) as Deployment;
 
       return Promise.resolve(response);
     } catch (error) {
-      return Promise.reject(error.response.data as NetworkErrorCollection);
-    }
+      const err = error as AxiosError;
+      if (err.response) {
+        return Promise.reject(err.response.data as NetworkErrorCollection);
+      }
+      return Promise.reject(err.message);    }
   }
 
   /**
@@ -100,11 +110,15 @@ export default class Deployments {
       const response = await this.client.call({
         endpoint: `apps/${appId}/deploys/${deployId}/logs`,
         method: "GET",
-      });
+      }) as string;
 
       return Promise.resolve(response);
     } catch (error) {
-      return Promise.reject(error.response.data as NetworkErrorCollection);
+      const err = error as AxiosError;
+      if (err.response) {
+        return Promise.reject(err.response.data as NetworkErrorCollection);
+      }
+      return Promise.reject(err.message);
     }
   }
 }
