@@ -1,12 +1,13 @@
 import Config from "./test-app/runner/config";
 import Runner from "./test-app/runner";
-import loadAndValidateApp from "./load-and-validate-app";
-import { 
-  CreateShipmentInternational, 
-  CreateShipmentDomestic, 
-  CreateShipmentWithInsurance, 
-  CreateShipmentMultiPackage,  
-  RateShipment
+import loadAndValidateApp, { isInvalidAppError } from "./load-and-validate-app";
+import {
+  CreateShipmentInternational,
+  CreateShipmentDomestic,
+  CreateShipmentWithInsurance,
+  CreateShipmentMultiPackage,
+  RateShipment,
+  CreateShipmentReturn
 } from "./test-app/tests";
 import { SdkApp } from "./types";
 import { TestResults, useTestResults } from "./test-app/runner/test-results";
@@ -14,6 +15,7 @@ import { loadAndValidateConfig } from "./test-app/runner/load-and-validate-confi
 import { logFail, logPass, logStep } from "./utils/log-helpers";
 import { logResults } from "./utils/log-helpers";
 import { RateShipmentWithAllServices } from './test-app/tests/rate-shipment-with-all-services';
+import { CancelShipment } from './test-app/tests/cancel-shipment';
 
 interface TesOptions {
   debug?: boolean;
@@ -29,6 +31,24 @@ export default async function testApp(
 ): Promise<TestResults> {
   const [testResults, testResultsReducer] = useTestResults();
 
+  // Set NODE_ENV first because its possible that the shipengine.config
+  // might key off the process.env to set environment variables
+  process.env.NODE_ENV = "test";
+
+  let staticConfig: Config = {};
+
+  try {
+    staticConfig = await loadAndValidateConfig(pathToApp);
+  } catch (error) {
+    switch (error.code) {
+      case "ERR_CONNECT_CONFIG_SCHEMA":
+        throw error;
+
+      default:
+        break;
+    }
+  }
+
   let app: SdkApp;
 
   try {
@@ -39,42 +59,28 @@ export default async function testApp(
     logPass("app structure is valid");
     testResultsReducer("INCREMENT_PASSED");
   } catch (error) {
-    switch (error.code) {
-      case "INVALID_APP":
-        // eslint-disable-next-line no-case-declarations
-        const errorsCount = error.errors.length;
-        // eslint-disable-next-line no-case-declarations
-        const errorsWithInflection = errorsCount > 1 ? "errors" : "error";
 
-        logFail(
-          `App structure is not valid - ${errorsCount} ${errorsWithInflection} found`,
-        );
+    if (isInvalidAppError(error)) {
+      const errorsCount = error.errors.length;
+      const errorsWithInflection = errorsCount > 1 ? "errors" : "error";
 
-        error.errors.forEach((errorMessage: string) => {
-          logFail(errorMessage);
-        });
+      logFail(
+        `App structure is not valid - ${errorsCount} ${errorsWithInflection} found`,
+      );
 
-        for (let i = 0; i < errorsCount; i++) {
-          testResultsReducer("INCREMENT_FAILED");
-        }
+      error.errors.forEach((errorMessage: string) => {
+        logFail(errorMessage);
+      });
 
-        logResults(testResults);
-        return testResults;
-      default:
-        throw error;
+      for (let i = 0; i < errorsCount; i++) {
+        testResultsReducer("INCREMENT_FAILED");
+      }
+
+      logResults(testResults);
+      return testResults;
     }
-  }
 
-  // Set NODE_ENV first because its possible that the shipengine.config
-  // might key off the process.env to set environment variables
-  process.env.NODE_ENV = "test";
-
-  let staticConfig: Config = {};
-
-  try {
-    staticConfig = await loadAndValidateConfig(pathToApp);
-  } catch {
-    // Do nothing
+    throw error;
   }
 
   const options = {
@@ -138,13 +144,15 @@ type RegisteredTestSuiteModules = object[];
 function registerTestSuiteModules(app: SdkApp): RegisteredTestSuiteModules {
   const carrierAppMethods = {
     // cancelPickups: [CancelPickupsTestSuite],
-    // cancelShipments: [CancelShipmentsTestSuite],
+    cancelShipments: [CancelShipment],
     // createManifest: [CreateManifestTestSuite],
     createShipment: [
-      CreateShipmentInternational, 
-      CreateShipmentDomestic, 
+      CreateShipmentInternational,
+      CreateShipmentDomestic,
       CreateShipmentMultiPackage,
-      CreateShipmentWithInsurance
+      CreateShipmentWithInsurance,
+      CreateShipmentWithInsurance,
+      CreateShipmentReturn
     ],
     rateShipment: [
       RateShipment,
