@@ -5,7 +5,7 @@ export default async function callWithTimeoutAndRetries(
   fn: any,
   timeout: number,
   retries: number,
-) {
+): Promise<unknown> {
   const run = async () => {
     return pTimeout(fn(), timeout, `test timed out after ${timeout}ms`);
   };
