@@ -7,13 +7,13 @@ import { BaseTestConfigOptions } from "./base-test-config-options";
 
 export interface RateShipmentTestParams {
   deliveryServiceName: string;
+  packagingName: string;
+  shipDateTime: DateTimeZonePOJO | Date | string;
   shipFrom: AddressWithContactInfoPOJO;
   shipTo: AddressWithContactInfoPOJO;
   weight: WeightPOJO;
-  shipDateTime: DateTimeZonePOJO | Date | string;
-  packagingName: string;
 }
 
 export interface RateShipmentConfigOptions
   extends RateShipmentTestParams,
-    BaseTestConfigOptions {}
+  BaseTestConfigOptions { }
