@@ -10,20 +10,21 @@ import {
   CreateShipmentWithInsuranceTestParams,
 } from "../runner/config/create-shipment-insurance";
 import { initializeTimeStamps } from "../../utils/time-stamps";
-import reduceDefaultsWithConfig from '../utils/reduce-defaults-with-config';
-import objectToTestTitle from '../utils/object-to-test-title';
-import useDomesticShippingAddress from '../utils/use-domestic-shipment-addresses';
+import reduceDefaultsWithConfig from "../utils/reduce-defaults-with-config";
+import objectToTestTitle from "../utils/object-to-test-title";
+import useDomesticShippingAddress from "../utils/use-domestic-shipment-addresses";
 
 import { expect } from "chai";
-import findDeliveryServiceByName from '../utils/find-delivery-service-by-name';
-import findDeliveryConfirmationByName from '../utils/find-delivery-confirmation-by-name';
-import { findInsurableDeliveryService } from '../utils/find-insurable-delivery-service';
-import findPackagingByName from '../utils/find-packaging-by-name';
+import findDeliveryServiceByName from "../utils/find-delivery-service-by-name";
+import findDeliveryConfirmationByName from "../utils/find-delivery-confirmation-by-name";
+import { findInsurableDeliveryService } from "../utils/find-insurable-delivery-service";
+import findPackagingByName from "../utils/find-packaging-by-name";
+import Test from '../runner/test';
 
 interface TestArgs {
   title: string;
   methodArgs: NewShipmentPOJO;
-  config: any;
+  config: unknown;
   testParams: CreateShipmentWithInsuranceTestParams;
 }
 
@@ -96,8 +97,8 @@ export class CreateShipmentWithInsurance extends Suite {
     const defaults: CreateShipmentWithInsuranceTestParams = {
       deliveryServiceName: this.deliveryService.name,
       shipDateTime: tomorrow,
-      shipFrom: shipFrom!,
-      shipTo: shipTo!,
+      shipFrom: shipFrom,
+      shipTo: shipTo,
       weight: {
         unit: WeightUnit.Pounds,
         value: 50.0,
@@ -136,8 +137,8 @@ export class CreateShipmentWithInsurance extends Suite {
       deliveryService: {
         id: this.deliveryService.id,
       },
-      shipFrom: testParams.shipFrom!,
-      shipTo: testParams.shipTo!,
+      shipFrom: testParams.shipFrom,
+      shipTo: testParams.shipTo,
       shipDateTime: testParams.shipDateTime,
       packages: [packagePOJO],
     };
@@ -149,11 +150,16 @@ export class CreateShipmentWithInsurance extends Suite {
     }
 
     if (testParams.deliveryConfirmationName) {
-      newShipmentPOJO.deliveryConfirmation = {
-        id: this.deliveryService.deliveryConfirmations.find(
-          (dc) => dc.name === testParams.deliveryConfirmationName,
-        )!.id,
-      };
+
+      const deliveryConfirmation = this.deliveryService.deliveryConfirmations.find(
+        (dc) => dc.name === testParams.deliveryConfirmationName
+      );
+
+      if (deliveryConfirmation) {
+        newShipmentPOJO.deliveryConfirmation = {
+          id: deliveryConfirmation.id,
+        };
+      }
     }
 
     const title = config.expectedErrorMessage
@@ -178,32 +184,32 @@ export class CreateShipmentWithInsurance extends Suite {
         return this.buildTestArg(config);
       });
     }
- 
+
     const config = this.config as CreateShipmentWithInsuranceConfigOptions;
     return [this.buildTestArg(config)];
   }
 
-  tests() {
-    const testArgs = this.buildTestArgs().filter((args) => args !== undefined);
+  tests(): Test[] {
+    const testArgs = this.buildTestArgs().filter((args) => args !== undefined) as TestArgs[];
 
     if (testArgs.length === 0) {
       return [];
     }
     return testArgs.map((testArg) => {
       return this.test(
-        testArg!.title,
-        testArg!.methodArgs,
-        testArg!.config,
+        testArg.title,
+        testArg.methodArgs,
+        testArg.config,
         async () => {
           const carrierApp = this.app as CarrierApp;
 
-          const transaction = await this.transaction(testArg!.config);
+          const transaction = await this.transaction(testArg.config);
 
           if (!carrierApp.createShipment) {
             throw new Error("createShipment is not implemented");
           }
 
-          const shipmentConfirmation = await carrierApp.createShipment(transaction, testArg!.methodArgs);
+          const shipmentConfirmation = await carrierApp.createShipment(transaction, testArg.methodArgs);
 
           // If DeliveryServiceDefinition.isTrackable is true, then the shipment must have a trackingNumber set
           if (this.deliveryService?.isTrackable) {
@@ -212,7 +218,7 @@ export class CreateShipmentWithInsurance extends Suite {
           }
 
           const customMsg = "The shipment confirmation packages array should have the same number of packages that were on the request";
-          expect(shipmentConfirmation.packages.length).to.equal(testArg!.methodArgs.packages.length, customMsg);
+          expect(shipmentConfirmation.packages.length).to.equal(testArg.methodArgs.packages.length, customMsg);
         }
       );
     });
