@@ -5,11 +5,11 @@ export * from "./create-shipment-domestic";
 export * from "./create-shipment-multipackage";
 export * from "./create-shipment-with-insurance";
 export * from "./create-shipment-international";
+export * from "./create-shipment-return";
 // export * from "./get-sales-order-test-suite";
 // export * from "./get-sales-orders-by-date-test-suite";
 // export * from "./get-seller-test-suite";
 export * from "./rate-shipment";
 // export * from "./schedule-pickup-test-suite";
-// export * from "./shipment-cancelled-test-suite";
 // export * from "./shipment-created-test-suite";
 // export * from "./track-shipment-test-suite";
