@@ -13,11 +13,12 @@ import reduceDefaultsWithConfig from "../utils/reduce-defaults-with-config";
 import objectToTestTitle from "../utils/object-to-test-title";
 import useShipmentAddresses from '../utils/use-shipment-addresses';
 import findDeliveryServiceByName from '../utils/find-delivery-service-by-name';
+import Test from '../runner/test';
 
 interface TestArgs {
   title: string;
   methodArgs: RateCriteriaPOJO;
-  config: any;
+  config: unknown;
   testParams: RateShipmentTestParams;
 }
 
@@ -83,7 +84,7 @@ export class RateShipment extends Suite {
     const RateCriteriaPOJO: RateCriteriaPOJO = {
       deliveryService: findDeliveryServiceByName(this.deliveryService.name, carrierApp),
       shipFrom: testParams.shipFrom,
-      shipTo: testParams.shipTo!,
+      shipTo: testParams.shipTo,
       shipDateTime: testParams.shipDateTime,
       packages: [packageRateCriteriaPOJO]
     };
@@ -115,7 +116,7 @@ export class RateShipment extends Suite {
     return [this.buildTestArg(config)];
   }
 
-  tests() {
+  tests(): Test[] {
     const testArgs = this.buildTestArgs().filter(args => args !== undefined) as TestArgs[];
 
     if (testArgs.length === 0) {
@@ -136,15 +137,15 @@ export class RateShipment extends Suite {
             throw new Error("rateShipment is not implemented");
           }
 
-          const rates = await carrierApp.rateShipment!(transaction, testArg!.methodArgs);
+          const rates = await carrierApp.rateShipment(transaction, testArg.methodArgs);
 
           // Check that the delivery service that was passed in is included in the response
           if (isDefinition(testArg.methodArgs.deliveryService)) {
-            const configuredDeliveryServiceID = Reflect.get(testArg.methodArgs.deliveryService, "id");
+            const configuredDeliveryServiceID = Reflect.get(testArg.methodArgs.deliveryService, "id") as string;
             if (!rates.find(rate => rate.deliveryService.id === configuredDeliveryServiceID)) {
-              const missingDS = carrierApp.deliveryServices.find(deliveryService => deliveryService.id === configuredDeliveryServiceID);
+              const missingDS = carrierApp.deliveryServices.find(deliveryService => deliveryService.id === configuredDeliveryServiceID) as DeliveryService;
 
-              throw new Error(`Rate for delivery service '${missingDS!.name}' is missing from the response`);
+              throw new Error(`Rate for delivery service '${missingDS.name}' is missing from the response`);
             }
           }
         }
@@ -154,5 +155,5 @@ export class RateShipment extends Suite {
 }
 
 function isDefinition(obj: unknown): obj is Definition {
-  return (typeof obj === "object") && ("id" in obj!);
+  return typeof obj === "object" && obj !== null && "id" in obj;
 }
