@@ -7,7 +7,7 @@ export enum DeploymentStatus {
   Error = "error",
 }
 
-export type Deployment = {
+export interface Deployment {
   package: {
     name: string;
     version: string | null;
@@ -16,4 +16,4 @@ export type Deployment = {
   status: DeploymentStatus;
   createdAt: string;
   updatedAt: string;
-};
+}
