@@ -2,7 +2,7 @@
 declare module "netrc" {
   interface NetRC {
     (): Record<string, object>;
-    save: (netRC: object) => void
+    save: (netRC: object) => void;
     parse: (netRC: string) => object;
     format: (netRC: object) => string;
   } 
