@@ -3,10 +3,10 @@
 export const flatRatePackaging: PackagingConfig = {
   id: "7c012ad2-71c3-11ea-bc55-0242ac130003",
   name: "Flat-Rate Box"
-}
+};
 
 /** Large Padded Envelope Packaging */
 export const largePaddedEnvelope: PackagingConfig = {
   id: "e7d6906a-72ba-11ea-bc55-0242ac130003",
   name: "Large Padded Envelope"
-}
+};
