@@ -1,30 +1,15 @@
-/* eslint-disable camelcase */
 "use strict";
 
-const { CreateShipmentWithInsurance } = require("../../../../../lib/core/test-app/tests/create-shipment-with-insurance");
+const { CancelShipment } = require("../../../../../lib/core/test-app/tests/cancel-shipment");
 const { CarrierApp } = require("@shipengine/connect-sdk/lib/internal/carriers/carrier-app");
 const pojo = require("../../../utils/pojo");
 const { expect } = require("chai");
 const sinon = require("sinon");
+const { v4 } = require("uuid");
 
-describe("The create shipment insured test suite", () => {
+describe("The cancel shipment test suite", () => {
 
-  describe("when there is no delivery service that supports insurance", () => {
-
-    it("should not generate tests", () => {
-      const { appDefinition, connectArgs, staticConfigTests, options } = generateBasicAppAndConfigs();
-      appDefinition.deliveryServices[0].isInsurable = false;
-      const app = new CarrierApp(appDefinition);
-      const args = { app, connectArgs, staticConfigTests, options };
-      const testSuite = new CreateShipmentWithInsurance(args);
-
-      const tests = testSuite.tests();
-      expect(tests.length).to.equal(0);
-    });
-  });
-
-
-  describe("when there is not address available for a delivery service services", () => {
+  describe("when there is not address available for the delivery services", () => {
     it("should not generate tests", () => {
       const { appDefinition, connectArgs, staticConfigTests, options } = generateBasicAppAndConfigs();
       appDefinition.deliveryServices[0].originCountries = ["AQ"];
@@ -32,14 +17,14 @@ describe("The create shipment insured test suite", () => {
 
       const app = new CarrierApp(appDefinition);
       const args = { app, connectArgs, staticConfigTests, options };
-      const testSuite = new CreateShipmentWithInsurance(args);
+      const testSuite = new CancelShipment(args);
 
       const tests = testSuite.tests();
       expect(tests.length).to.equal(0);
     });
   });
 
-  describe("when there is a insurable delivery service with an available address", () => {
+  describe("when there is a delivery service with an available address", () => {
 
     let testSuite;
     beforeEach(() => {
@@ -47,7 +32,7 @@ describe("The create shipment insured test suite", () => {
       const app = new CarrierApp(appDefinition);
       const args = { app, connectArgs, staticConfigTests, options };
 
-      testSuite = new CreateShipmentWithInsurance(args);
+      testSuite = new CancelShipment(args);
     });
 
     it("should generate a test", () => {
@@ -58,7 +43,6 @@ describe("The create shipment insured test suite", () => {
     it("the test params should be reflected in the title", () => {
       const tests = testSuite.tests();
 
-      expect(tests[0].title).to.include("packageInsuredValue: 10 USD");
       expect(tests[0].title).to.include("weight: 50lb");
     });
   });
@@ -73,25 +57,18 @@ describe("The create shipment insured test suite", () => {
 
       const app = new CarrierApp(appDefinition);
 
-      staticConfigTests.createShipment_with_insurance = {
+      staticConfigTests.cancelShipment = {
         weight: {
           value: 200,
           unit: "lb"
-        },
-        packagingName: "New Package",
-        packageInsuredValue: {
-          value: "50",
-          currency: "usd"
         }
       };
 
       const args = { app, connectArgs, staticConfigTests, options };
-      const testSuite = new CreateShipmentWithInsurance(args);
+      const testSuite = new CancelShipment(args);
       const tests = testSuite.tests();
 
-      expect(tests[0].title).to.include("packagingName: New Package");
-      expect(tests[0].title).to.include("packageInsuredValue: 50 usd");
-
+      expect(tests[0].title).to.include("weight: 200lb");
     });
   });
 
@@ -101,32 +78,24 @@ describe("The create shipment insured test suite", () => {
     beforeEach(() => {
       const { appDefinition, connectArgs, staticConfigTests, options } = generateBasicAppAndConfigs();
       const app = new CarrierApp(appDefinition);
-      staticConfigTests.createShipment_with_insurance =
+      staticConfigTests.cancelShipment =
         [
           {
             weight: {
               value: 200,
               unit: "lb"
-            },
-            packageInsuredValue: {
-              value: "50",
-              currency: "usd"
             }
           },
           {
             weight: {
               value: 22,
               unit: "lb"
-            },
-            packageInsuredValue: {
-              value: "50",
-              currency: "usd"
             }
           }
         ];
 
       const args = { app, connectArgs, staticConfigTests, options };
-      const testSuite = new CreateShipmentWithInsurance(args);
+      const testSuite = new CancelShipment(args);
       tests = testSuite.tests();
     });
 
@@ -137,11 +106,8 @@ describe("The create shipment insured test suite", () => {
 
     it("should update the test titles", () => {
       expect(tests[0].title).to.include("weight: 200lb");
-      expect(tests[0].title).to.include("packageInsuredValue: 50 usd");
-
 
       expect(tests[1].title).to.include("weight: 22lb");
-      expect(tests[1].title).to.include("packageInsuredValue: 50 usd");
     });
   });
 
@@ -149,12 +115,12 @@ describe("The create shipment insured test suite", () => {
     it("should throw an error", () => {
       const { appDefinition, connectArgs, staticConfigTests, options } = generateBasicAppAndConfigs();
       const app = new CarrierApp(appDefinition);
-      staticConfigTests.createShipment_with_insurance = {
+      staticConfigTests.cancelShipment = {
         deliveryServiceName: "asdf"
       };
 
       const args = { app, connectArgs, staticConfigTests, options };
-      const testSuite = new CreateShipmentWithInsurance(args);
+      const testSuite = new CancelShipment(args);
 
       try {
         testSuite.tests();
@@ -166,52 +132,16 @@ describe("The create shipment insured test suite", () => {
     });
   });
 
-  describe("When a user configs a delivery service that is not insurable", () => {
-    it("should throw an error", () => {
-      const { appDefinition, connectArgs, staticConfigTests, options } = generateBasicAppAndConfigs();
-      appDefinition.deliveryServices.push({
-        id: "9cf1bfda-7ee4-4f03-96f6-6eab52243eee",
-        isInsurable: false,
-        name: "Uninsured Delivery Service",
-        code: "uninsured_ds",
-        class: "ground",
-        manifestType: "digital",
-        grade: "standard",
-        originCountries: ["MX"],
-        destinationCountries: ["MX"],
-        labelFormats: ["pdf"],
-        labelSizes: ["A4"],
-        packaging: [pojo.packaging()]
-      });
-
-      const app = new CarrierApp(appDefinition);
-      staticConfigTests.createShipment_with_insurance = {
-        deliveryServiceName: "Uninsured Delivery Service"
-      };
-      const args = { app, connectArgs, staticConfigTests, options };
-      const testSuite = new CreateShipmentWithInsurance(args);
-
-      try {
-        testSuite.tests();
-        expect(true).to.equal(false);
-      }
-      catch (error) {
-        expect(error.message).to.include("The configured delivery service 'Uninsured Delivery Service' does not support insuring packages");
-      }
-    });
-  });
-
   describe("When a user configs a new delivery service", () => {
     it("should update the title params to reflect the new properties", () => {
       const { appDefinition, connectArgs, staticConfigTests, options } = generateBasicAppAndConfigs();
       appDefinition.deliveryServices.push({
         id: "9cf1bfda-7ee4-4f03-96f6-6eab52243eee",
-        isInsurable: true,
         name: "Better Delivery Service",
-        manifestType: "digital",
         class: "ground",
-        code: "priority_overnight",
+        code: "better_ds",
         grade: "standard",
+        manifestType: "physical",
         originCountries: ["MX"],
         destinationCountries: ["MX"],
         labelFormats: ["pdf"],
@@ -219,26 +149,30 @@ describe("The create shipment insured test suite", () => {
         packaging: [pojo.packaging()]
       });
 
-      staticConfigTests.createShipment_with_insurance = {
+      staticConfigTests.cancelShipment = {
         deliveryServiceName: "Better Delivery Service"
-      }
+      };
 
       const app = new CarrierApp(appDefinition);
       const args = { app, connectArgs, staticConfigTests, options };
-      const testSuite = new CreateShipmentWithInsurance(args);
+      const testSuite = new CancelShipment(args);
       const tests = testSuite.tests();
 
       expect(tests[0].title).to.include("deliveryServiceName: Better Delivery Service");
     });
   });
 
-  describe("When a user configures a Ship To and Ship From address", () => {
-    it("should update the test arguments and titles", () => {
+  describe("When a delivery service has addresses that we don't have samples but user uses valid configs", () => {
+    it("should generate tests", () => {
       const { appDefinition, connectArgs, staticConfigTests, options } = generateBasicAppAndConfigs();
+
+      appDefinition.deliveryServices[0].originCountries = ["AQ", "US"];
+      appDefinition.deliveryServices[0].destinationCountries = ["AQ", "US"];
+
 
       const app = new CarrierApp(appDefinition);
 
-      staticConfigTests.createShipment_with_insurance = {
+      staticConfigTests.cancelShipment = {
         shipFrom: {
           company: "Domestic Route #1",
           addressLines: ["123 New Street"],
@@ -260,13 +194,45 @@ describe("The create shipment insured test suite", () => {
       };
 
       const args = { app, connectArgs, staticConfigTests, options };
-      const testSuite = new CreateShipmentWithInsurance(args);
+      const testSuite = new CancelShipment(args);
+      const tests = testSuite.tests();
+      expect(tests.length).to.equal(1);
+    });
+  });
+
+  describe("When a user configures a Ship To and Ship From address", () => {
+    it("should update the test arguments and titles", () => {
+      const { appDefinition, connectArgs, staticConfigTests, options } = generateBasicAppAndConfigs();
+
+      const app = new CarrierApp(appDefinition);
+
+      staticConfigTests.cancelShipment = {
+        shipFrom: {
+          company: "Domestic Route #1",
+          addressLines: ["123 New Street"],
+          cityLocality: "Houston",
+          stateProvince: "TX",
+          country: "US",
+          postalCode: "77422"
+        },
+        shipTo: {
+          company: "Domestic Route #2",
+          addressLines: ["123 New Street"],
+          cityLocality: "Houston",
+          stateProvince: "TX",
+          country: "US",
+          postalCode: "77422"
+        }
+      };
+
+      const args = { app, connectArgs, staticConfigTests, options };
+      const testSuite = new CancelShipment(args);
       const tests = testSuite.tests();
 
       expect(tests[0].methodArgs.shipFrom.company).to.equal("Domestic Route #1");
       expect(tests[0].methodArgs.shipTo.company).to.equal("Domestic Route #2");
 
-      expect(tests[0].methodArgs.shipTo).to.eql(staticConfigTests.createShipment_with_insurance.shipTo);
+      expect(tests[0].methodArgs.shipTo).to.eql(staticConfigTests.cancelShipment.shipTo);
 
       expect(tests[0].title).to.include("shipFrom: US");
       expect(tests[0].title).to.include("shipTo: US");
@@ -274,73 +240,56 @@ describe("The create shipment insured test suite", () => {
     });
   });
 
-  describe("When a delivery service 'isTrackable' property is set", () => {
-    it("should throw an error if no tracking number is returned", async () => {
-      const { appDefinition, connectArgs, staticConfigTests, options } = generateBasicAppAndConfigs();
+  describe("When a valid shipment is created and the cancellation ids do not match", () => {
 
-      appDefinition.deliveryServices[0].isTrackable = true;
-      const confirmationMock = pojo.shipmentConfirmation();
-      sinon.stub(CarrierApp.prototype, "createShipment").resolves(confirmationMock);
-      const app = new CarrierApp(appDefinition);
-
-      const args = { app, connectArgs, staticConfigTests, options };
-      const testSuite = new CreateShipmentWithInsurance(args);
-      const tests = testSuite.tests();
-      try {
-        await tests[0].fn();
-        expect(true).to.equal(false);
-      }
-      catch (error) {
-        expect(error.message).includes("The shipmentConfirmation.isTrackable returned from createShipment must be present when the given deliveryService.isTrackable is set to 'true'");
-      }
-    });
-
-    afterEach(() => {
-      CarrierApp.prototype.createShipment.restore();
-    });
-  });
-
-  describe("When the input parameters do not match the return shipment", () => {
-
-    it("should throw an error for a packaging length mismatch", async () => {
+    it("should throw an error", async () => {
       const { appDefinition, connectArgs, staticConfigTests, options } = generateBasicAppAndConfigs();
 
       const confirmationMock = pojo.shipmentConfirmation();
       confirmationMock.packages.push(pojo.packageConfirmation());
-      sinon.stub(CarrierApp.prototype, "createShipment").resolves(confirmationMock);
-      const app = new CarrierApp(appDefinition);
 
+      sinon.stub(CarrierApp.prototype, "createShipment").resolves(confirmationMock);
+
+      const shipmentCancelledOutcomeMock = [{
+        cancellationId: v4()
+      }];
+
+      sinon.stub(CarrierApp.prototype, "cancelShipments").resolves(shipmentCancelledOutcomeMock);
+
+      const app = new CarrierApp(appDefinition);
       const args = { app, connectArgs, staticConfigTests, options };
-      const testSuite = new CreateShipmentWithInsurance(args);
+      const testSuite = new CancelShipment(args);
       const tests = testSuite.tests();
+
       try {
         await tests[0].fn();
         expect(true).to.equal(false);
       }
       catch (error) {
-        expect(error.message).includes("The shipment confirmation packages array should have the same number of packages that were on the request");
+        expect(error.message).includes("The shipmentCancellationConfirmation cancellationID does not match the one that was included in the shipmentCancellation");
       }
     });
 
     afterEach(() => {
       CarrierApp.prototype.createShipment.restore();
+      CarrierApp.prototype.cancelShipments.restore();
     });
-
   });
+
 });
 
 function generateBasicAppAndConfigs() {
   const appDefinition = pojo.carrierApp();
   const deliveryService = pojo.deliveryService();
-  deliveryService.labelFormats = ["pdf"];
   deliveryService.manifestType = "digital";
-  deliveryService.labelSizes = ["A4"];
-  deliveryService.isInsurable = true;
+  deliveryService.labelFormats = ["pdf"];
   deliveryService.code = "priority_overnight";
+  deliveryService.labelSizes = ["A4"];
   deliveryService.deliveryConfirmations = [pojo.deliveryConfirmation()];
   deliveryService.packaging.push(pojo.packaging());
   appDefinition.deliveryServices = [deliveryService];
   appDefinition.createShipment = () => { };
+  appDefinition.cancelShipments = () => { };
 
   const options = {
     cli: {
@@ -358,7 +307,7 @@ function generateBasicAppAndConfigs() {
   };
 
   const staticConfigTests = {
-    createShipment_with_insurance: {}
+    cancelShipment: {}
   };
 
   const connectArgs = {};
