@@ -1,22 +1,23 @@
 /* eslint-disable camelcase */
 "use strict";
 
-const { CreateShipmentWithInsurance } = require("../../../../../lib/core/test-app/tests/create-shipment-with-insurance");
+const { CreateShipmentReturn } = require("../../../../../lib/core/test-app/tests/create-shipment-return");
 const { CarrierApp } = require("@shipengine/connect-sdk/lib/internal/carriers/carrier-app");
 const pojo = require("../../../utils/pojo");
 const { expect } = require("chai");
 const sinon = require("sinon");
 
-describe("The create shipment insured test suite", () => {
+describe("The create shipment return test suite", () => {
 
-  describe("when there is no delivery service that supports insurance", () => {
+  describe("when there is no delivery service", () => {
 
     it("should not generate tests", () => {
       const { appDefinition, connectArgs, staticConfigTests, options } = generateBasicAppAndConfigs();
-      appDefinition.deliveryServices[0].isInsurable = false;
+      appDefinition.deliveryServices[0].originCountries = ["MX"];
+
       const app = new CarrierApp(appDefinition);
       const args = { app, connectArgs, staticConfigTests, options };
-      const testSuite = new CreateShipmentWithInsurance(args);
+      const testSuite = new CreateShipmentReturn(args);
 
       const tests = testSuite.tests();
       expect(tests.length).to.equal(0);
@@ -24,7 +25,7 @@ describe("The create shipment insured test suite", () => {
   });
 
 
-  describe("when there is not address available for a delivery service services", () => {
+  describe("when there is not address available for a delivery service", () => {
     it("should not generate tests", () => {
       const { appDefinition, connectArgs, staticConfigTests, options } = generateBasicAppAndConfigs();
       appDefinition.deliveryServices[0].originCountries = ["AQ"];
@@ -32,14 +33,14 @@ describe("The create shipment insured test suite", () => {
 
       const app = new CarrierApp(appDefinition);
       const args = { app, connectArgs, staticConfigTests, options };
-      const testSuite = new CreateShipmentWithInsurance(args);
+      const testSuite = new CreateShipmentReturn(args);
 
       const tests = testSuite.tests();
       expect(tests.length).to.equal(0);
     });
   });
 
-  describe("when there is a insurable delivery service with an available address", () => {
+  describe("when there is a delivery service with an available address", () => {
 
     let testSuite;
     beforeEach(() => {
@@ -47,7 +48,7 @@ describe("The create shipment insured test suite", () => {
       const app = new CarrierApp(appDefinition);
       const args = { app, connectArgs, staticConfigTests, options };
 
-      testSuite = new CreateShipmentWithInsurance(args);
+      testSuite = new CreateShipmentReturn(args);
     });
 
     it("should generate a test", () => {
@@ -58,7 +59,7 @@ describe("The create shipment insured test suite", () => {
     it("the test params should be reflected in the title", () => {
       const tests = testSuite.tests();
 
-      expect(tests[0].title).to.include("packageInsuredValue: 10 USD");
+      expect(tests[0].title).to.include("label: A4 pdf");
       expect(tests[0].title).to.include("weight: 50lb");
     });
   });
@@ -73,25 +74,24 @@ describe("The create shipment insured test suite", () => {
 
       const app = new CarrierApp(appDefinition);
 
-      staticConfigTests.createShipment_with_insurance = {
+      staticConfigTests.createShipment_return = {
         weight: {
           value: 200,
           unit: "lb"
         },
-        packagingName: "New Package",
-        packageInsuredValue: {
-          value: "50",
-          currency: "usd"
+
+        label: {
+          size: "A6",
+          format: "png"
         }
       };
 
       const args = { app, connectArgs, staticConfigTests, options };
-      const testSuite = new CreateShipmentWithInsurance(args);
+      const testSuite = new CreateShipmentReturn(args);
       const tests = testSuite.tests();
 
-      expect(tests[0].title).to.include("packagingName: New Package");
-      expect(tests[0].title).to.include("packageInsuredValue: 50 usd");
-
+      expect(tests[0].title).to.include("label: A6 png");
+      expect(tests[0].title).to.include("weight: 200lb");
     });
   });
 
@@ -101,32 +101,28 @@ describe("The create shipment insured test suite", () => {
     beforeEach(() => {
       const { appDefinition, connectArgs, staticConfigTests, options } = generateBasicAppAndConfigs();
       const app = new CarrierApp(appDefinition);
-      staticConfigTests.createShipment_with_insurance =
+      staticConfigTests.createShipment_return =
         [
           {
             weight: {
               value: 200,
               unit: "lb"
             },
-            packageInsuredValue: {
-              value: "50",
-              currency: "usd"
+            label: {
+              size: "A6",
+              format: "png"
             }
           },
           {
             weight: {
               value: 22,
               unit: "lb"
-            },
-            packageInsuredValue: {
-              value: "50",
-              currency: "usd"
             }
           }
         ];
 
       const args = { app, connectArgs, staticConfigTests, options };
-      const testSuite = new CreateShipmentWithInsurance(args);
+      const testSuite = new CreateShipmentReturn(args);
       tests = testSuite.tests();
     });
 
@@ -137,11 +133,9 @@ describe("The create shipment insured test suite", () => {
 
     it("should update the test titles", () => {
       expect(tests[0].title).to.include("weight: 200lb");
-      expect(tests[0].title).to.include("packageInsuredValue: 50 usd");
-
+      expect(tests[0].title).to.include("label: A6 png");
 
       expect(tests[1].title).to.include("weight: 22lb");
-      expect(tests[1].title).to.include("packageInsuredValue: 50 usd");
     });
   });
 
@@ -149,12 +143,12 @@ describe("The create shipment insured test suite", () => {
     it("should throw an error", () => {
       const { appDefinition, connectArgs, staticConfigTests, options } = generateBasicAppAndConfigs();
       const app = new CarrierApp(appDefinition);
-      staticConfigTests.createShipment_with_insurance = {
+      staticConfigTests.createShipment_return = {
         deliveryServiceName: "asdf"
-      };
+      }
 
       const args = { app, connectArgs, staticConfigTests, options };
-      const testSuite = new CreateShipmentWithInsurance(args);
+      const testSuite = new CreateShipmentReturn(args);
 
       try {
         testSuite.tests();
@@ -166,52 +160,16 @@ describe("The create shipment insured test suite", () => {
     });
   });
 
-  describe("When a user configs a delivery service that is not insurable", () => {
-    it("should throw an error", () => {
-      const { appDefinition, connectArgs, staticConfigTests, options } = generateBasicAppAndConfigs();
-      appDefinition.deliveryServices.push({
-        id: "9cf1bfda-7ee4-4f03-96f6-6eab52243eee",
-        isInsurable: false,
-        name: "Uninsured Delivery Service",
-        code: "uninsured_ds",
-        class: "ground",
-        manifestType: "digital",
-        grade: "standard",
-        originCountries: ["MX"],
-        destinationCountries: ["MX"],
-        labelFormats: ["pdf"],
-        labelSizes: ["A4"],
-        packaging: [pojo.packaging()]
-      });
-
-      const app = new CarrierApp(appDefinition);
-      staticConfigTests.createShipment_with_insurance = {
-        deliveryServiceName: "Uninsured Delivery Service"
-      };
-      const args = { app, connectArgs, staticConfigTests, options };
-      const testSuite = new CreateShipmentWithInsurance(args);
-
-      try {
-        testSuite.tests();
-        expect(true).to.equal(false);
-      }
-      catch (error) {
-        expect(error.message).to.include("The configured delivery service 'Uninsured Delivery Service' does not support insuring packages");
-      }
-    });
-  });
-
   describe("When a user configs a new delivery service", () => {
     it("should update the title params to reflect the new properties", () => {
       const { appDefinition, connectArgs, staticConfigTests, options } = generateBasicAppAndConfigs();
       appDefinition.deliveryServices.push({
         id: "9cf1bfda-7ee4-4f03-96f6-6eab52243eee",
-        isInsurable: true,
         name: "Better Delivery Service",
-        manifestType: "digital",
         class: "ground",
-        code: "priority_overnight",
+        code: "better_ds",
         grade: "standard",
+        manifestType: "physical",
         originCountries: ["MX"],
         destinationCountries: ["MX"],
         labelFormats: ["pdf"],
@@ -219,26 +177,31 @@ describe("The create shipment insured test suite", () => {
         packaging: [pojo.packaging()]
       });
 
-      staticConfigTests.createShipment_with_insurance = {
+      staticConfigTests.createShipment_return = {
         deliveryServiceName: "Better Delivery Service"
       }
 
       const app = new CarrierApp(appDefinition);
       const args = { app, connectArgs, staticConfigTests, options };
-      const testSuite = new CreateShipmentWithInsurance(args);
+      const testSuite = new CreateShipmentReturn(args);
       const tests = testSuite.tests();
 
       expect(tests[0].title).to.include("deliveryServiceName: Better Delivery Service");
+      expect(tests[0].title).to.include("label: A4 pdf");
     });
   });
 
-  describe("When a user configures a Ship To and Ship From address", () => {
-    it("should update the test arguments and titles", () => {
+  describe("When a delivery service has addresses that we don't have samples but user uses valid configs", () => {
+    it("should generate tests", () => {
       const { appDefinition, connectArgs, staticConfigTests, options } = generateBasicAppAndConfigs();
+
+      appDefinition.deliveryServices[0].originCountries = ["AQ", "US"];
+      appDefinition.deliveryServices[0].destinationCountries = ["AQ", "US"];
+
 
       const app = new CarrierApp(appDefinition);
 
-      staticConfigTests.createShipment_with_insurance = {
+      staticConfigTests.createShipment_return = {
         shipFrom: {
           company: "Domestic Route #1",
           addressLines: ["123 New Street"],
@@ -248,7 +211,7 @@ describe("The create shipment insured test suite", () => {
           postalCode: "77422",
           timeZone: "America/Chicago"
         },
-        shipTo: {
+        returnTo: {
           company: "Domestic Route #2",
           addressLines: ["123 New Street"],
           cityLocality: "Houston",
@@ -260,16 +223,48 @@ describe("The create shipment insured test suite", () => {
       };
 
       const args = { app, connectArgs, staticConfigTests, options };
-      const testSuite = new CreateShipmentWithInsurance(args);
+      const testSuite = new CreateShipmentReturn(args);
+      const tests = testSuite.tests();
+      expect(tests.length).to.equal(1);
+    });
+  });
+
+  describe("When a user configures a Ship To and Ship From address", () => {
+    it("should update the test arguments and titles", () => {
+      const { appDefinition, connectArgs, staticConfigTests, options } = generateBasicAppAndConfigs();
+
+      const app = new CarrierApp(appDefinition);
+
+      staticConfigTests.createShipment_return = {
+        shipFrom: {
+          company: "Domestic Route #1",
+          addressLines: ["123 New Street"],
+          cityLocality: "Houston",
+          stateProvince: "TX",
+          country: "US",
+          postalCode: "77422"
+        },
+        returnTo: {
+          company: "Domestic Route #2",
+          addressLines: ["123 New Street"],
+          cityLocality: "Houston",
+          stateProvince: "TX",
+          country: "US",
+          postalCode: "77422"
+        }
+      };
+
+      const args = { app, connectArgs, staticConfigTests, options };
+      const testSuite = new CreateShipmentReturn(args);
       const tests = testSuite.tests();
 
       expect(tests[0].methodArgs.shipFrom.company).to.equal("Domestic Route #1");
       expect(tests[0].methodArgs.shipTo.company).to.equal("Domestic Route #2");
 
-      expect(tests[0].methodArgs.shipTo).to.eql(staticConfigTests.createShipment_with_insurance.shipTo);
+      expect(tests[0].methodArgs.shipTo).to.eql(staticConfigTests.createShipment_return.returnTo);
 
       expect(tests[0].title).to.include("shipFrom: US");
-      expect(tests[0].title).to.include("shipTo: US");
+      expect(tests[0].title).to.include("returnTo: US");
 
     });
   });
@@ -284,7 +279,7 @@ describe("The create shipment insured test suite", () => {
       const app = new CarrierApp(appDefinition);
 
       const args = { app, connectArgs, staticConfigTests, options };
-      const testSuite = new CreateShipmentWithInsurance(args);
+      const testSuite = new CreateShipmentReturn(args);
       const tests = testSuite.tests();
       try {
         await tests[0].fn();
@@ -311,7 +306,7 @@ describe("The create shipment insured test suite", () => {
       const app = new CarrierApp(appDefinition);
 
       const args = { app, connectArgs, staticConfigTests, options };
-      const testSuite = new CreateShipmentWithInsurance(args);
+      const testSuite = new CreateShipmentReturn(args);
       const tests = testSuite.tests();
       try {
         await tests[0].fn();
@@ -332,11 +327,10 @@ describe("The create shipment insured test suite", () => {
 function generateBasicAppAndConfigs() {
   const appDefinition = pojo.carrierApp();
   const deliveryService = pojo.deliveryService();
-  deliveryService.labelFormats = ["pdf"];
   deliveryService.manifestType = "digital";
-  deliveryService.labelSizes = ["A4"];
-  deliveryService.isInsurable = true;
+  deliveryService.labelFormats = ["pdf"];
   deliveryService.code = "priority_overnight";
+  deliveryService.labelSizes = ["A4"];
   deliveryService.deliveryConfirmations = [pojo.deliveryConfirmation()];
   deliveryService.packaging.push(pojo.packaging());
   appDefinition.deliveryServices = [deliveryService];
@@ -358,7 +352,7 @@ function generateBasicAppAndConfigs() {
   };
 
   const staticConfigTests = {
-    createShipment_with_insurance: {}
+    createShipment_return: {}
   };
 
   const connectArgs = {};
